made `Node` an actual class in the module; added example #2
`Node` is not just a class found in the example anymore; now it's an actual class from the module. It also has a class method to process dictionaries of `str => dict` items, in which each `str` is the node name and its respective `dict` is its child nodes. I also added an example of how this method should be used.